+++ packages/components/layout/README.md
@@ -1,0 +1,3 @@
+# @contentful/f36-layout
+
+This package is part of [Forma-36](https://github.com/contentful/forma-36). See the repo for more details.
