+++ packages/components/layout/src/Grid/Grid.tsx
@@ -4,7 +4,7 @@ import {
   PolymorphicComponentProps,
   PolymorphicComponent,
   Primitive,
-} from '../Primitive';
+} from '@contentful/f36-core';
 import { cx, css } from 'emotion';
 import type * as CSS from 'csstype';
