+++ packages/core/src/index.ts
@@ -1,3 +1,2 @@
 export * from './Primitive';
-export * from './Grid';
 export type { CommonProps } from './types';
